move fossils and stones to their own data files
